requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
decorator==5.1.1
discord==2.2.3
discord.py==2.2.3
fakeyou==1.2.5
frozenlist==1.3.3
idna==3.4
inworld-python==0.0.10
joblib==1.2.0
lazy_loader==0.2
librosa==0.10.0.post2
llvmlite==0.40.0
msgpack==1.0.5
multidict==6.0.4
nodejs-bin==18.4.0a4
nodejs-cmd==0.0.1a0
numba==0.57.0
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pooch==1.6.0
pycparser==2.21
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
soundfile==0.12.1
soxr==0.3.5
threadpoolctl==3.1.0
typing_extensions==4.6.3
urllib3==2.0.2
yarl==1.9.2
